requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-WTF
email_validator
flask_login
mysql-connector-python